refactor query tests, place in query folder
