+++ server/package.json
@@ -8,10 +8,11 @@
   "main": "index.js",
   "scripts": {
     "start": "nodemon ./bin/app",
+    "database:run:dev": "docker run -d -p 27017-27019:27017-27019 --name mongodb mongo",
     "database:build:dev": "babel-node ./database/data",
     "database:build:fake": "babel-node ./database/data && npm run fake",
     "fake": "babel-node ./database/data/fakedata",
-    "test": "echo \"No test specified\"",
+    "test": "jest --runInBand --detectOpenHandles --forceExit --watch",
     "test:coverage": "echo \"No test specified\""
   },
   "dependencies": {
@@ -21,6 +22,7 @@
     "dotenv": "^8.1.0",
     "express": "^4.17.1",
     "http-errors": "^1.7.3",
+    "jest": "^24.9.0",
     "mongoose": "^5.7.4",
     "pino": "^5.13.4",
     "pino-http": "^4.3.0",
